+++ Software/src/services/communication/fleshy.hpp
@@ -14,11 +14,11 @@ class FTSCallbacks : public NimBLECharacteristicCallbacks {
         // position: uint8 (0-180), convert to 100
         // time: uint16 big-endian (MSB first)
         if (value.length() >= 3) {
-            uint8_t position = static_cast<uint8_t>(value[0] / 1.8);
+            float position = static_cast<float>(value[0] / 1.8);
             uint16_t time = (static_cast<uint8_t>(value[1]) << 8) |
                             static_cast<uint8_t>(value[2]);
 
-            ESP_LOGI("NIMBLE", "FTS Command - Position: %d, Time: %d ms", position, time);
+            ESP_LOGI("NIMBLE", "FTS Command - Position: %f, Time: %d ms", position, time);
             targetQueue.push({position, time, std::chrono::steady_clock::now(), 0});
 
         } else {
