+++ Software/src/services/communication/queue.h
@@ -6,7 +6,7 @@
 #include <queue>
 
 struct PositionTime {
-    uint8_t position; // 0 - 100
+    float position; // 0 - 100
     uint16_t inTime;     // in ms
     std::chrono::steady_clock::time_point setTime; //received timestamp
     int direction; //0:uncalculated, 1:out, -1:in
